refactor(common): move schema ifaces to `schema` package
